Conflicts:
	src/Library/Helper/Directory.php
	src/Library/Helper/Url.php

+++ src/Library/Helper/Directory.php
@@ -24,6 +24,16 @@ class Directory
 {
 
     /**
+     * @var int
+     */
+    const DEFAULT_UNIX_CHMOD_DIRECTORIES = 755;
+
+    /**
+     * @var int
+     */
+    const DEFAULT_UNIX_CHMOD_FILES = 644;
+
+    /**
      * Get a dirname with one and only trailing slash
      *
      * @param string $dirname
@@ -60,7 +70,7 @@ class Directory
      * @param bool $recursive
      * @return bool
      */
-    public static function ensureExists($path, $mode = 777, $recursive = true)
+    public static function ensureExists($path, $mode = self::DEFAULT_UNIX_CHMOD_DIRECTORIES, $recursive = true)
     {
         if (file_exists($path) && is_dir($path)) return true;
         return self::create($path, $mode, $recursive);
@@ -74,7 +84,7 @@ class Directory
      * @param bool $recursive
      * @return bool
      */
-    public static function create($path, $mode = 777, $recursive = true)
+    public static function create($path, $mode = self::DEFAULT_UNIX_CHMOD_DIRECTORIES, $recursive = true)
     {
         return mkdir($path, Filesystem::getOctal($mode), $recursive);
     }
@@ -229,35 +239,45 @@ class Directory
                 if (unlink($path)) {
                     return true;
                 } else {
-                    $logs[] = sprintf('Can not unlink file "%s".', $path);
+                    $logs[$path] = sprintf('Can not unlink file "%s".', $path);
                 }
             }
             $iterator = new \RecursiveIteratorIterator(
                 new \RecursiveDirectoryIterator($path), 
                 \RecursiveIteratorIterator::SELF_FIRST | \FilesystemIterator::CURRENT_AS_FILEINFO | \FilesystemIterator::SKIP_DOTS
             );
-            foreach($iterator as $item) {
+            foreach ($iterator as $item) {
+                if (in_array($item->getFilename(), array('.', '..'))) {
+                    continue;
+                }
+                $_path = $item->getRealpath();
+                if (array_key_exists($_path, $logs)) {
+                    return false;
+                }
                 if ($item->isDir()) {
-                    if (false===$ok = self::remove($item, $logs)) {
-                        $logs[] = sprintf('Can not remove diretory "%s".', $item);
+                    if (false===$ok = self::remove($_path, $logs)) {
+                        $logs[$_path] = sprintf('Can not remove diretory "%s".', $_path);
                     }
                 } else {
-                    if (false===$ok = File::remove($item, $logs)) {
-                        $logs[] = sprintf('Can not unlink file "%s".', $item);
+                    if (false===$ok = File::remove($_path, $logs)) {
+                        $logs[$_path] = sprintf('Can not unlink file "%s".', $_path);
                     }
                 }
             } 
             if (true===$ok) {
+                if (array_key_exists($path, $logs)) {
+                    return false;
+                }
                 if (rmdir($path)) {
                     return true;
                 } else {
-                    $logs[] = sprintf('Can not remove directory "%s".', $path);
+                    $logs[$path] = sprintf('Can not remove directory "%s".', $path);
                 }
             }
             clearstatcache();
             return $ok;
         } else {
-            $logs[] = sprintf('Directory "%s" not found.', $path);
+            $logs[$path] = sprintf('Directory "%s" not found.', $path);
         }
         return false;
     }
@@ -272,8 +292,10 @@ class Directory
      * @param array $logs Logs registry passed by reference
      * @return bool
      */
-    public static function chmod($path, $mode = 777, $recursive = true, $file_mode = 777, array &$logs = array())
-    {
+    public static function chmod(
+        $path, $mode = self::DEFAULT_UNIX_CHMOD_DIRECTORIES,
+        $recursive = true, $file_mode = self::DEFAULT_UNIX_CHMOD_FILES, array &$logs = array()
+    ){
         $ok = false;
         if (file_exists($path) && is_dir($path)) {
             if (true!==$ok = chmod($path, Filesystem::getOctal($mode))) {
@@ -285,6 +307,9 @@ class Directory
                     \RecursiveIteratorIterator::SELF_FIRST | \FilesystemIterator::CURRENT_AS_FILEINFO | \FilesystemIterator::SKIP_DOTS
                 );
                 foreach($iterator as $item) {
+                    if (in_array($item->getFilename(), array('.', '..'))) {
+                        continue;
+                    }
                     if ($item->isDir()) {
                         if (true!==$ok = chmod($item, Filesystem::getOctal($mode))) {
                             $logs[] = sprintf('Can not change mode on sub-directory "%s" (trying to set them on "%d").', $item, $mode);
